moved input validations to forms util
